+++ Classes/GHBlob.h
@@ -14,5 +14,6 @@
 @property(nonatomic,strong)NSData *contentData;
 @property(nonatomic,assign)NSUInteger size;
 
-- (id)initWithRepo:(GHRepository *)repo andSha:(NSString *)sha;
+- (id)initWithRepo:(GHRepository *)repo path:(NSString *)path ref:(NSString*)ref;
+- (id)initWithRepo:(GHRepository *)repo sha:(NSString *)sha;
 @end