+++ Classes/IOCBlobsController.h
@@ -1,3 +1,6 @@
+@class GHBlob;
+
 @interface IOCBlobsController : UIViewController
 - (id)initWithBlobs:(NSArray *)blobs currentIndex:(NSUInteger)idx;
+- (id)initWithBlob:(GHBlob *)blob;
 @end